==> code/step_2_working_data.rmd <==
---
title: "Modify Raw Imported Data"
author: "Jonathan Salas"
date: "Last Updated `r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---

# Start with a clean slate

## Remove all objects

```{r echo = F, include=FALSE}
rm(list = ls(all.names = TRUE)) # Remove all starting code
```

## Detach Packages

```{r include=FALSE}
try(invisible(lapply(paste0('package:', names(sessionInfo()$otherPkgs)), detach, character.only=TRUE, unload=TRUE)), silent = T) # Detach Packages
```

# Load Libraries

```{r setup, include = FALSE}
library(stringr)
library(dplyr)
knitr::opts_knit$set(root.dir = "~/R_Projects/Dose_Response")
```

# Import the raw data - Point to the correct raw data directory

```{r}
raw_data_dir <- "data/processed/imported_dr.RData"
```

# Import Processed VCAS and PULSE Labs Data

```{r}
load("~/R_Projects/VCAS1_prelim_labs/data/processed/working_stats.RData")
VCAS_hemolysis_perc_change <- hemolysis_perc_change
load("~/R_Projects/PULSE_prelim_labs/data/processed/working_stats.RData")
PULSE_hemolysis_perc_change <- hemolysis_perc_change

rm(cleaned_res, cleaned_res_1, headers, hemolysis_dat, hemolysis_perc_change)
```

## Import Subject Data
```{r}
load(raw_data_dir)
```

# Combine Hemolysis Percent Change 

```{r}
combined_hemolysis_perc_change <- 
  bind_rows(VCAS_hemolysis_perc_change |> mutate(study = "VCAS") |> rename(sum_applications = numberapplications),
      PULSE_hemolysis_perc_change |> mutate(study = "PULSE")) 
```

## Correct Data Types & Calculate derived statistics

```{r}
cleaned_res <- cleaned_res |>
  mutate(
    id = factor(id),
    lesion_num = as.numeric(lesion_num),
    n_uni = as.numeric(n_uni),
    n_bipolar = as.numeric(n_bipolar),
    n_biphasic = as.numeric(n_biphasic),
    n_train_rep = as.numeric(n_train_rep),
    depth = as.numeric(depth),
    diameter = as.numeric(diameter),
    # Derived stats
    depth_per_width = depth/diameter,
    # Rounding
    across(where(is.numeric), \(x) round(x, 2))
  ) 
```

```{r}
cleaned_res_2 <- cleaned_res_2 |>
  mutate(
    id = factor(id),
    lesion_num = as.numeric(lesion_num),
    n_uni = as.numeric(n_uni),
    n_bipolar = as.numeric(n_bipolar),
    n_biphasic = as.numeric(n_biphasic),
    n_train_rep = as.numeric(n_train_rep),
    dimension = factor(dimension),
    dim_value = as.numeric(dim_value),
    transmural = factor(transmural),
    # Derived stats

    # Rounding
    across(where(is.numeric), \(x) round(x, 2))
  ) 
```


# Make Analysis Data
 
We exclude the fast-switch data from the 04022024Potato experiment because here we varied ratio of bipolar to unipolar applications to assess depth & diameter. Isn't directly comparable to the bulk of the data. We thus exclude the 10x5 bipolar 04022024Potato data (n=30)

We also exclude the four 20x5 bipolar 15kV lesions from "3x12x3 potato tests_08_02_2024" per Roya's advise they were done with a different catheter which at this stage adds too much variability to the lesion data. (n=4)

At some point in August 2024, someone changed the file name from "3x12x3 potato tests_08_02_202" to "unipolar-3x12x3-08_02_2024"

We will exclude the non-3x12x3 data from the "unipolar-3x12x3-08_02_2024" experiment in  working_dat and include it in full_curve_3x12x3


```{r}
working_dat <- cleaned_res |>
  filter(!(experiment_id == "04022024Potato" & pulse_seq == "10x5" )) |> 
  filter(!(experiment_id == "unipolar-3x12x3-08_02_2024" & pulse_seq == "20x5")) |>
  filter(!(experiment_id == "unipolar-3x12x3-08_02_2024" & pulse_seq != "3x12x3"))
```

```{r}
temp <- cleaned_res |> anti_join(y = working_dat)
```

```{r}
full_curve_3x12x3 <- cleaned_res |>
  filter(experiment_id == "unipolar-3x12x3-08_02_2024" & pulse_seq != "20x5") |>
  mutate(pulses = n_uni*n_train_rep*application_num) |>
  arrange(pulses)
```

## Deduplicate transmural depths

When transmural, we mark both rows as having the same depth, but that's deceptive

```{r}
kissing_dat <-  cleaned_res_2 |> distinct()
```

# Save cleaned data

```{r}
save(working_dat, 
     full_curve_3x12x3,
     kissing_dat,
     file = paste0("data/processed/working_dr.RData"))
```

```{r}
save(combined_hemolysis_perc_change, 
     file = paste0("data/processed/combined_hemolysis_dat.RData"))
```


# Version and Package Details

```{r echo = F}
paste(version$version.string, version$nickname)
paste0(
  "RStudio Version ",
  rstudioapi::versionInfo()$version,
  " ",
  rstudioapi::versionInfo()$release_name
) # RStudio Version
subset(data.frame(sessioninfo::package_info()),
       attached == TRUE,
       c(package, loadedversion)) # Loaded Packages
```

# When were these files last rewritten?

```{r}
date()
```
