==> code/step_3_descriptive_stats.rmd <==
---
title: "Dose-Response Descriptive Statistics" 
author: "Jonathan Salas"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document: 
    df_print: kable
    fig_height: 6
    keep_md: true
---

```{r setup, include=FALSE}
library(knitr)
library(janitor)
library(ggplot2)
library(cowplot)
library(lme4)
library(lmerTest)
library(dplyr)
knitr::opts_chunk$set(echo = F)
knitr::opts_knit$set(root.dir = "~/R_Projects/Dose_Response")
```

# Start with a clean slate

```{r echo = F}
rm(list = ls(all.names = TRUE))
```

# Import the cleaned data - Point to the correct raw data directory

```{r echo = F}
working_dat_dir <- "data/processed/working_dr.RData"
```

## Import Subject Data
```{r echo = F}
load(working_dat_dir)
```

# Descriptive Statistics

## Summary Table of Subjects 
```{r How many subjects got which waveform, pulse sequence, and input voltage, echo = F}
working_dat |>
  count(species, waveform, voltage, pulse_seq) |> left_join(
    # Add in basic stats for depth/diameter
    working_dat |> group_by(species, waveform, voltage, pulse_seq) |>
      summarise(
        median_depth = median(depth, na.rm = T),
        IQR_depth = IQR(depth, na.rm = T),
        median_diameter = median(diameter, na.rm = T),
        IQR_diameter = IQR(diameter, na.rm = T),
        median_depth_per_width = median(depth_per_width, na.rm = T),
        IQR_depth_per_width = IQR(depth_per_width, na.rm = T)
      )
  )
```

```{r}
kissing_dat |>
  filter(dimension == "depth") |>
  count(species, waveform, voltage, pulse_seq, application_num) |> left_join(
    # Add in basic stats for depth/diameter
    kissing_dat |>
      filter(dimension == "depth") |> group_by(species, waveform, voltage, pulse_seq, application_num) |>
      summarise(
        median_depth = median(dim_value, na.rm = T),
        IQR_depth = IQR(dim_value, na.rm = T)
      )
  )
```

```{r}
kissing_dat |>
  filter(dimension == "diameter") |>
  count(species, waveform, voltage, pulse_seq, application_num) |> left_join(
    # Add in basic stats for depth/diameter
    kissing_dat |>
      filter(dimension == "diameter") |> group_by(species, waveform, voltage, pulse_seq, application_num) |>
      summarise(
        median_diameter = median(dim_value, na.rm = T),
        IQR_diameter = IQR(dim_value, na.rm = T)
      )
  )
```

# Plots

## Depth

### Depth Dose-Response - Line Plots

```{r echo = F}
working_dat |>
  ggplot(data = _, aes(x = application_num, y = depth)) +
  geom_point(
    size = 4,
    na.rm = T,
    aes(color = interaction(voltage, waveform, pulse_seq)),
    position = position_jitter(seed = 42494)
  ) +
  facet_wrap(~ voltage + waveform + pulse_seq, nrow = 1) +
  scale_x_continuous(labels = working_dat$application_num,
                     breaks = working_dat$application_num) +
  scale_color_discrete(guide = "none") +
  coord_cartesian(ylim = c(0, 25)) +
  #  ggtitle("Title") +
  xlab("Applications") +
  ylab("Depth (mm)") +
  # labs(caption = "Caption") +
  theme_cowplot(12) +
  theme(
    axis.text = element_text(size = 30, face = "bold"),
    axis.title = element_text(size = 24, face = "bold")
  ) 

ggsave("raw_applications_vs_depth.png",
       path = "figures",
       scale = 1,
       width = 60,
       height = 25,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

### 15kV 3x12x3 Unipolar Full Depth Curve

```{r}
#temp <- colour_ramp(brewer_pal(
#          palette = "Spectral", direction = -1
#        )(11))(rescale(full_curve_3x12x3$pulses))
```

```{r}
full_curve_3x12x3 |>
  ggplot(data = _, aes(x = pulses, y = depth, color = factor(pulses))) +
  geom_point(size = 4, na.rm = T, position = position_jitter(seed = 42494)) +
  scale_x_continuous(
#    labels = factor(full_curve_3x12x3$pulse_seq),
#    breaks = full_curve_3x12x3$pulses,
    transform = "log2"
  ) +
  #  coord_cartesian(ylim = c(0, 16)) +
  #  ggtitle("Title") +
  xlab("Pulses") +
  ylab("Depth (mm)") +
  # labs(caption = "Caption") +
  theme_cowplot(12)
```

### 10kV 20x5 Kissing Catheters - 1 vs 3 applications

```{r}
kissing_dat |>
  filter(dimension == "depth") |>
  ggplot(data = _,
         aes(
           x = factor(application_num),
           y = dim_value,
           group = factor(application_num),
           color = transmural,
           shape = id
         )) +
  geom_boxplot() +
  geom_point(size = 4, na.rm = T, position = position_jitter(seed = 42494)) +
  #  scale_x_continuous(
  #    labels = factor(full_curve_3x12x3$pulse_seq),
  #    breaks = full_curve_3x12x3$pulses,
  #    trans = "log2"
  #  ) +
  #  coord_cartesian(ylim = c(0, 16)) +
  ggtitle("10kV 20x5 Kissing Catheters") +
  xlab("Applications") +
  ylab("Depth (mm)") +
  # labs(caption = "Caption") +
  theme_cowplot(12)
```

#### Depth Dose-Response - Smoothed Line Plots

```{r echo = F}
working_dat |>
  ggplot(data = _, aes(x = application_num, y = depth)) +
  geom_smooth(span = 10) +
  geom_point(size = 4, na.rm = T, position = position_jitter(seed = 42494)) +
  facet_wrap( ~ voltage + waveform + pulse_seq, nrow = 1) +
  scale_x_continuous(labels = working_dat$application_num,
                     breaks = working_dat$application_num) +
  scale_color_discrete(guide = "none") +
  coord_cartesian(ylim = c(0,25)) +
#  ggtitle("Title") +
  xlab("Applications") +
  ylab("Depth (mm)") +
# labs(caption = "Caption") +
  theme_cowplot(12)
```

## Diameter Dose-Response

### Diameter line plots 

```{r}
working_dat |>
  ggplot(data = _, aes(x = application_num, y = diameter)) +
  geom_point(
    size = 4,
    na.rm = T,
    aes(color = interaction(voltage, waveform, pulse_seq)),
    position = position_jitter(seed = 42494)
  ) +
facet_wrap(~ voltage + waveform + pulse_seq, nrow = 1) +
  scale_x_continuous(labels = working_dat$application_num,
                     breaks = working_dat$application_num) +
  scale_color_discrete(guide = "none") +
  coord_cartesian(ylim = c(0, 60)) +
  #  ggtitle("Title") +
  xlab("Applications") +
  ylab("Diameter (mm)") +
  # labs(caption = "Caption") +
  theme_cowplot(12) +
  theme(
    axis.text = element_text(size = 30, face = "bold"),
    axis.title = element_text(size = 24, face = "bold")
  ) 

ggsave("raw_applications_vs_diameter.png",
       path = "figures",
       scale = 1,
       width = 60,
       height = 25,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```
### 15kV 3x12x3 Unipolar Full Diameter Curve

```{r}
full_curve_3x12x3 |>
  ggplot(data = _, aes(x = pulses, y = diameter)) +
  geom_point(size = 4, na.rm = T, position = position_jitter(seed = 42494)) +
  scale_x_continuous(
    labels = factor(full_curve_3x12x3$pulse_seq),
    breaks = full_curve_3x12x3$pulses,
    trans = "log2"
  ) +
  #  coord_cartesian(ylim = c(0, 16)) +
  #  ggtitle("Title") +
  xlab("Pulses") +
  ylab("Depth (mm)") +
  # labs(caption = "Caption") +
  theme_cowplot(12)
```

### 10kV 20x5 Kissing Catheters - 1 vs 3 applications

```{r}
kissing_dat |>
  filter(dimension == "diameter") |>
  ggplot(data = _,
         aes(
           x = factor(application_num),
           y = dim_value,
           group = factor(application_num),
           color = transmural,
           shape = id
         )) +
  geom_boxplot(outliers=F) +
  geom_point(size = 4, na.rm = T, position = position_jitter(seed = 42494)) +
  #  scale_x_continuous(
  #    labels = factor(full_curve_3x12x3$pulse_seq),
  #    breaks = full_curve_3x12x3$pulses,
  #    trans = "log2"
  #  ) +
  #  coord_cartesian(ylim = c(0, 16)) +
  ggtitle("10kV 20x5 Kissing Catheters") +
  xlab("Applications") +
  ylab("Diameter (mm)") +
  # labs(caption = "Caption") +
  theme_cowplot(12)
```

#### Depth Dose-Response - Smoothed Line Plots

```{r echo = F}
working_dat |>
  ggplot(data = _, aes(x = application_num, y = depth)) +
  geom_smooth(span = 10) +
  geom_point(size = 4, na.rm = T, position = position_jitter(seed = 42494)) +
  facet_wrap( ~ voltage + waveform + pulse_seq, nrow = 1) +
  scale_x_continuous(labels = working_dat$application_num,
                     breaks = working_dat$application_num) +
  scale_color_discrete(guide = "none") +
  coord_cartesian(ylim = c(0,25)) +
#  ggtitle("Title") +
  xlab("Applications") +
  ylab("Depth (mm)") +
# labs(caption = "Caption") +
  theme_cowplot(12)
```



#### Diameter Dose-Response - Smoothed Line Plots

```{r}
working_dat |>
  ggplot(data = _, aes(x = application_num, y = diameter)) +
  geom_smooth(span = 10) +
  geom_point(size = 4, na.rm = T, position = position_jitter(seed = 42494)) +
  facet_wrap( ~ voltage + waveform + pulse_seq, nrow = 1) +
  scale_x_continuous(labels = working_dat$application_num,
                     breaks = working_dat$application_num) +
  scale_color_discrete(guide = "none") +
 # ggtitle("Title") +
  xlab("Applications") +
  ylab("Diameter (mm)") +
# labs(caption = "Caption") +
  theme_cowplot(12)
```

## Depth/Diameter Ratio

### Depth/Diameter Ratio line plots 

```{r}
working_dat |>
  ggplot(data = _, aes(x = application_num, y = depth_per_width)) +
  geom_point(
    size = 4,
    na.rm = T,
    aes(color = interaction(voltage, waveform, pulse_seq)),
    position = position_jitter(seed = 42494)
  ) +
  facet_wrap(~ voltage + waveform + pulse_seq, nrow = 1) +
  scale_x_continuous(labels = working_dat$application_num,
                     breaks = working_dat$application_num) +
  scale_color_discrete(guide = "none") +
  #  ggtitle("Title") +
  xlab("Applications") +
  ylab("Depth/Diameter Ratio") +
  # labs(caption = "Caption") +
  theme_cowplot(12) +
  theme(
    axis.text = element_text(size = 30, face = "bold"),
    axis.title = element_text(size = 24, face = "bold")
  ) 

ggsave("raw_depth_diameter_ratio.png",
       path = "figures",
       scale = 1,
       width = 60,
       height = 25,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```


### Depth/Diameter Ratio LOESS line plots 

```{r}
working_dat |>
  ggplot(data = _, aes(x = application_num, y = depth_per_width)) +
  geom_point(size = 4,
             na.rm = T,
             position = position_jitter(seed = 42494)) +
  geom_smooth(span = 10) +
  facet_wrap(~ voltage + waveform + pulse_seq, nrow = 1) +
  scale_x_continuous(labels = working_dat$application_num,
                     breaks = working_dat$application_num) +
  scale_color_discrete(guide = "none") +
  # ggtitle("Title") +
  xlab("Applications") +
  ylab("Depth/Diameter Ratio") +
  # labs(caption = "Caption") +
  theme_cowplot(12)
```


# Version and Package Details

```{r echo=F}
paste(version$version.string, version$nickname)
paste0(
  "RStudio Version ",
  rstudioapi::versionInfo()$version,
  " ",
  rstudioapi::versionInfo()$release_name
) # RStudio Version
subset(data.frame(sessioninfo::package_info()),
       attached == TRUE,
       c(package, loadedversion)) # Loaded Packages
```

# When were these files last rewritten?

```{r}
date()
```
