==> code/step_4_regression.rmd <==
---
title: "Regression Modeling" 
author: "Jonathan Salas"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document: 
    df_print: kable
    fig_height: 6
    keep_md: true
---

```{r setup, include=FALSE}
library(knitr)
library(sfsmisc)
library(emmeans)
library(aomisc)
library(ggplot2)
library(dplyr)
knitr::opts_chunk$set(echo = F)
knitr::opts_knit$set(root.dir = "~/R_Projects/Dose_Response")
```

# Start with a clean slate

```{r echo = F}
rm(list = ls(all.names = TRUE))
```

# Import the cleaned data - Point to the correct raw data directory

```{r echo = F}
working_dat_dir <-"data/processed/working_dr.RData"
```

# Aesthetics

```{r echo = F}
blank.plot <- theme_bw() +
  theme(panel.border = element_blank(), 
        panel.grid.major = element_blank(), #this theme section generates a blank canvas
        panel.grid.minor = element_blank(), 
        axis.line = element_line(colour = "black"),
        legend.title = element_blank(),
        legend.justification="top")
```

## Import Subject Data
```{r echo = F}
load(working_dat_dir)
```

# Regression Fitting

The mselect method failed us for moseling. It wanted me to fit the 10kv 20x5 Bipolar diameter with a 3 parameter log-logistic curve. It should be linear by looking at it. 

## 10kV 20x5 Bipolar 

### Depth 

#### Model Selection 

```{r}
drm_depth_mselect_1 <- drm(
  depth ~ application_num,
  data = working_dat,
  subset = (voltage == 10 &
              pulse_seq == "20x5" & 
              waveform == "bipolar"),
  fct = AR.2(names = c("Upper Limit", "Steepness"))
)
```

```{r}
mselect(
  drm_depth_mselect_1,
  list(
    L.3(), #L.3 lower limit at dose 0 is fixed at 0
    LL.3(), #LL.3 lower limit at dose 0 is fixed at 0
    LL2.3(), #LL2.3 lower limit at dose 0 is fixed at 0
    AR.2(), #AR.2 lower limit at dose 0 is fixed at 0
    MM.2(), #MM.2 lower limit at dose = 0 is fixed at 0
  ),
  linreg = TRUE
)
```

#### Robust Linear Regression

```{r}
rlm_depth_1 <- rlm(
  depth ~ application_num,
  data = working_dat,
  subset = (voltage == 10 &
              pulse_seq == "20x5" &
              waveform == "bipolar")
)
```

```{r}
summary(rlm_depth_1)
```

##### Post-Hoc Tests

```{r}
f.robftest(rlm_depth_1, var = "application_num")
```

#### Plotting

```{r}
working_dat |>
  filter(voltage == 10,
              pulse_seq == "20x5",
              waveform == "bipolar") |>
ggplot(data = _, aes(x = application_num, y = depth)) +
  geom_jitter(width = 0.2, size = 4, na.rm = T) + 
  geom_smooth(method = "rlm", se = F ) +
  scale_x_continuous(labels = unique(working_dat$application_num),
                     breaks = unique(working_dat$application_num)) +
  ggtitle("rlm 10kV 20x5 Bipolar") +
  xlab("Applications") +
  ylab("Depth (mm)") +
# labs(caption = "Caption") +
  coord_cartesian(ylim=c(3,7)) +
  blank.plot +
  theme(axis.text=element_text(size=40, face = "bold"),
        axis.title=element_text(size=24,face="bold")) 

ggsave("10kv20x5bipolar_depth.png",
       path = "figures",
       scale = 1,
       width = 20,
       height = 28,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

### Diameter 

#### Model Selection 

```{r}
drm_diameter_mselect_1 <- drm(
  diameter ~ application_num,
  data = working_dat,
  subset = (voltage == 10 &
              pulse_seq == "20x5" & 
              waveform == "bipolar"),
  fct = AR.2(names = c("Upper Limit", "Steepness"))
)
```

```{r}
mselect(
  drm_diameter_mselect_1,
  list(
    L.3(), #L.3 lower limit at dose 0 is fixed at 0
    LL.3(), #LL.3 lower limit at dose 0 is fixed at 0
    LL2.3(), #LL2.3 lower limit at dose 0 is fixed at 0
    AR.2(), #AR.2 lower limit at dose 0 is fixed at 0
    MM.2(), #MM.2 lower limit at dose = 0 is fixed at 0
  ),
  linreg = TRUE
)
```

#### Robust Linear Regression

```{r}
rlm_diameter_1 <- rlm(
  diameter ~ application_num,
  data = working_dat,
  subset = (voltage == 10 &
              pulse_seq == "20x5" &
              waveform == "bipolar")
)

drm_diameter_1 <- drm(
  diameter ~ application_num,
  data = working_dat,
  subset = (voltage == 10 &
              pulse_seq == "20x5" &
              waveform == "bipolar"),
  fct = LL.3()
)
```

```{r}
summary(rlm_diameter_1)
summary(drm_diameter_1)
```

###### Post-Hoc Tests

```{r}
f.robftest(rlm_diameter_1, var = "application_num")
```

#### Plotting

```{r}
working_dat |>
  filter(voltage == 10,
              pulse_seq == "20x5",
              waveform == "bipolar") |>
ggplot(data = _, aes(x = application_num, y = diameter)) +
  geom_jitter(width = 0.2, size = 4, na.rm = T) + 
  geom_smooth(method = "rlm", se = F ) +
  scale_x_continuous(labels = unique(working_dat$application_num),
                     breaks = unique(working_dat$application_num)) +
  ggtitle("rlm 10kV 20x5 Bipolar") +
  xlab("Applications") +
  ylab("Diameter (mm)") +
# labs(caption = "Caption") +
# coord_cartesian(ylim=c(3,7)) +
  blank.plot +
  theme(axis.text=element_text(size=40, face = "bold"),
        axis.title=element_text(size=24,face="bold")) 

ggsave("10kv20x5bipolar_diameter.png",
       path = "figures",
       scale = 1,
       width = 20,
       height = 28,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```


### Depth/Diameter Ratio

#### Robust Linear Regression

```{r}
rlm_depth_ratio_1 <- rlm(
  depth_per_width ~ application_num,
  data = working_dat,
  subset = (voltage == 10 &
              pulse_seq == "20x5" & 
              waveform == "bipolar"))
```

##### Post-Hoc

```{r}
summary(rlm_depth_ratio_1)

f.robftest(rlm_depth_ratio_1, var = "application_num")
```

#### Plotting 

```{r}
working_dat |>
  filter(voltage == 10,
              pulse_seq == "20x5",
              waveform == "bipolar") |>
ggplot(data = _, aes(x = application_num, y = depth_per_width)) +
  geom_jitter(width = 0.2, size = 4, na.rm = T) + 
  geom_smooth(method = "rlm", se = F) +
  scale_x_continuous(labels = unique(working_dat$application_num),
                     breaks = unique(working_dat$application_num)) +
  ggtitle("rlm 10kV 20x5 Bipolar") +
  xlab("Applications") +
  ylab("Depth per Diameter") +
# labs(caption = "Caption") +
#  coord_cartesian(ylim=c(10,20)) +
  blank.plot +
  theme(axis.text=element_text(size=40, face = "bold"),
        axis.title=element_text(size=24,face="bold")) 
```

## 10 kV bipolar 10x5

### Depth

```{r}
drm_depth_asymp_2 <- drm(
  depth ~ application_num,
  data = working_dat,
  subset = (voltage == 10 &
              pulse_seq == "10x5" & 
              waveform == "bipolar"),
  fct = AR.2(names = c("Upper Limit", "Steepness"))
)
```

#### Model Selection

```{r}
mselect(
  drm_depth_asymp_2,
  list(
    L.3(), #L.3 lower limit at dose 0 is fixed at 0
    LL.3(), #LL.3 lower limit at dose 0 is fixed at 0
    LL2.3(), #LL2.3 lower limit at dose 0 is fixed at 0
    AR.2(), #AR.2 lower limit at dose 0 is fixed at 0
    MM.2(), #MM.2 lower limit at dose = 0 is fixed at 0
  ),
  linreg = TRUE
)
```

#### 3-Parameter Nonlinear Logistic Regression

```{r}
drm_depth_G3_2 <- drm(
  depth ~ application_num,
  data = working_dat,
  subset = (voltage == 10 &
              pulse_seq == "10x5" & 
              waveform == "bipolar"),
  fct = G.3(names = c("b","Upper","e")))
```

##### Post-Hoc

```{r}
summary(drm_depth_G3_2)
```

#### Plotting 

```{r}
working_dat |>
  filter(voltage == 10, pulse_seq == "10x5", waveform == "bipolar") |>
  ggplot(data = _, aes(x = application_num, y = depth)) +
  geom_smooth(method = drm,
              method.args = list(fct = G.3()),
              se = F) +
  geom_jitter(width = 0.2, size = 4, na.rm = T) + 
  scale_x_continuous(
    labels = unique(working_dat$application_num),
    breaks = unique(working_dat$application_num)
  ) +
  ggtitle("3-Parameter Gompertz 10kV 10x5 bipolar") +
  xlab("Applications") +
  ylab("Depth (mm)") +
  # labs(caption = "Caption") + 
  coord_cartesian(ylim=c(0,15))+
  blank.plot +
  theme(axis.text=element_text(size=40, face = "bold"),
        axis.title=element_text(size=24,face="bold")) 

ggsave("10kv10x5fast_switch_depth.png",
       path = "figures",
       scale = 1,
       width = 20,
       height = 28,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

### Diameter

#### Model Selection

```{r}
drm_diameter_asymp_2 <- drm(
  diameter ~ application_num,
  data = working_dat,
  subset = (voltage == 10 &
              pulse_seq == "10x5" & 
              waveform == "bipolar"),
  fct = AR.2(names = c("Upper Limit", "Steepness"))
)
```

```{r}
mselect(
  drm_diameter_asymp_2,
  list(
    L.3(), #L.3 lower limit at dose 0 is fixed at 0
    LL.3(), #LL.3 lower limit at dose 0 is fixed at 0
    LL2.3(), #LL2.3 lower limit at dose 0 is fixed at 0
    AR.2(), #AR.2 lower limit at dose 0 is fixed at 0
    MM.2(), #MM.2 lower limit at dose = 0 is fixed at 0
  ),
  linreg = TRUE
)
```

#### Four-Parameter Gompertz

```{r}
drm_diameter_G4_2 <- drm(
  diameter ~ application_num,
  data = working_dat,
  subset = (voltage == 10 &
              pulse_seq == "10x5" & 
              waveform == "bipolar"),
  fct = G.4())
```

```{r}
summary(drm_diameter_G4_2)
```

```{r}
working_dat |>
  filter(voltage == 10, pulse_seq == "10x5", waveform == "bipolar") |>
  ggplot(data = _, aes(x = application_num, y = diameter)) +
  geom_smooth(method = drm,
              method.args = list(fct = AR.2()),
              se = F) +
  geom_jitter(width = 0.2, size = 4, na.rm = T) + 
  scale_x_continuous(
    labels = unique(working_dat$application_num),
    breaks = unique(working_dat$application_num)
  ) +
  ggtitle("10kV 10x5 bipolar") +
  xlab("Applications") +
  ylab("Diameter (mm)") +
  # labs(caption = "Caption") + 
  coord_cartesian(ylim=c(15,25))+
  blank.plot +
  theme(axis.text=element_text(size=40, face = "bold"),
        axis.title=element_text(size=24,face="bold")) 

ggsave("10kv10x5fast_switch_diameter.png",
       path = "figures",
       scale = 1,
       width = 20,
       height = 28,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

#### Depth/Diameter Ratio

```{r}
rlm_depth_ratio_2 <- rlm(
  depth_per_width ~ application_num,
  data = working_dat,
  subset = (voltage == 10 &
              pulse_seq == "10x5" & 
              waveform == "bipolar"))
```

```{r}
summary(rlm_depth_ratio_2)
f.robftest(rlm_depth_ratio_2, var = "application_num")
```

```{r}
working_dat |>
  filter(voltage == 10,
              pulse_seq == "10x5",
              waveform == "bipolar") |>
ggplot(data = _, aes(x = application_num, y = depth_per_width)) +
  geom_jitter(width = 0.2, size = 4, na.rm = T) + 
  geom_smooth(method = "rlm", se = T) +
  scale_x_continuous(labels = unique(working_dat$application_num),
                     breaks = unique(working_dat$application_num)) +
  ggtitle("10kV 20x5 Bipolar") +
  xlab("Applications") +
  ylab("Depth per Diameter") +
  # labs(caption = "Caption") +
#  coord_cartesian(ylim=c(10,20)) +
  blank.plot 
```

## 15kV Unipolar 15x5

### 2-Parameter Nonlinear Asymptotic Regression

#### Depth

```{r}
drm_depth_asymp_3 <- drm(
  depth ~ application_num,
  data = working_dat,
  subset = (voltage == 15 &
              pulse_seq == "15x5" & 
              waveform == "unipolar"),
  fct = AR.2(names = c("Upper Limit", "Steepness"))
)
```

##### Model Selection 

```{r}
mselect(
  drm_depth_asymp_3,
  list(
    L.3(), #L.3 lower limit at dose 0 is fixed at 0
    LL.3(), #LL.3 lower limit at dose 0 is fixed at 0
    LL2.3(), #LL2.3 lower limit at dose 0 is fixed at 0
    AR.2(), #AR.2 lower limit at dose 0 is fixed at 0
    MM.2(), #MM.2 lower limit at dose = 0 is fixed at 0
  ),
  linreg = TRUE
)
```

```{r}
rlm_depth_asymp_3 <- rlm(
  depth ~ application_num,
  data = working_dat,
  subset = (voltage == 15 &
              pulse_seq == "15x5" & 
              waveform == "unipolar")
)
```

```{r}
summary(drm_depth_asymp_3)
summary(rlm_depth_asymp_3)
```

```{r}
working_dat |>
  filter(voltage == 15, pulse_seq == "15x5", waveform == "unipolar") |>
  ggplot(data = _, aes(x = application_num, y = depth)) +
  geom_smooth(method = drm,
              method.args = list(fct = AR.2()),
              se = F) +
  geom_jitter(width = 0.2, size = 4, na.rm = T) + 
  scale_x_continuous(
    labels = unique(working_dat$application_num),
    breaks = unique(working_dat$application_num)
  ) +
  ggtitle("15kV 15x5 Unipolar") +
  xlab("Applications") +
  ylab("Depth (mm)") +
  # labs(caption = "Caption") + 
  coord_cartesian(ylim=c(5,25))+
  blank.plot +
  theme(axis.text=element_text(size=40, face = "bold"),
        axis.title=element_text(size=24,face="bold")) 

ggsave("15kv15x5uni_depth.png",
       path = "figures",
       scale = 1,
       width = 20,
       height = 28,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

#### Diameter

```{r}
drm_diameter_asymp_3 <- drm(
  diameter ~ application_num,
  data = working_dat,
  subset = (voltage == 15 &
              pulse_seq == "15x5" & 
              waveform == "unipolar"),
  fct = AR.2(names = c("Upper Limit", "Steepness"))
)
```

##### Model Selection 

```{r}
mselect(
  drm_diameter_asymp_3,
  list(
    L.3(), #L.3 lower limit at dose 0 is fixed at 0
    LL.3(), #LL.3 lower limit at dose 0 is fixed at 0
    LL2.3(), #LL2.3 lower limit at dose 0 is fixed at 0
    AR.2(), #AR.2 lower limit at dose 0 is fixed at 0
    MM.2(), #MM.2 lower limit at dose = 0 is fixed at 0
  ),
  linreg = TRUE
)
```

```{r}
summary(drm_diameter_asymp_3)
```

```{r}
working_dat |>
  filter(voltage == 15, pulse_seq == "15x5", waveform == "unipolar") |>
  ggplot(data = _, aes(x = application_num, y = diameter)) +
  geom_smooth(method = drm,
              method.args = list(fct = AR.2()),
              se = F) +
  geom_jitter(width = 0.2, size = 4, na.rm = T) + 
  scale_x_continuous(
    labels = unique(working_dat$application_num),
    breaks = unique(working_dat$application_num)
  ) +
  ggtitle("15kV 15x5 Unipolar") +
  xlab("Applications") +
  ylab("Diameter (mm)") +
  # labs(caption = "Caption") + 
  coord_cartesian(ylim=c(20,50))+
  blank.plot +
  theme(axis.text=element_text(size=40, face = "bold"),
        axis.title=element_text(size=24,face="bold")) 

ggsave("15kv15x5uni_diameter.png",
       path = "figures",
       scale = 1,
       width = 20,
       height = 28,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

#### Depth/Diameter Ratio

```{r}
rlm_depth_ratio_3 <- rlm(
  depth_per_width ~ application_num,
  data = working_dat,
  subset = (voltage == 15 &
              pulse_seq == "15x5" & 
              waveform == "unipolar"))
```

```{r}
summary(rlm_depth_ratio_3)
f.robftest(rlm_depth_ratio_3, var = "application_num")
```

```{r}
working_dat |>
  filter(voltage == 15,
              pulse_seq == "15x5",
              waveform == "unipolar") |>
ggplot(data = _, aes(x = application_num, y = depth_per_width)) +
  geom_jitter(width = 0.2, size = 4, na.rm = T) + 
  geom_smooth(method = "rlm", se = T) +
  scale_x_continuous(labels = unique(working_dat$application_num),
                     breaks = unique(working_dat$application_num)) +
  ggtitle("15kV 15x5 Unipolar") +
  xlab("Applications") +
  ylab("Depth per Diameter") +
  # labs(caption = "Caption") +
#  coord_cartesian(ylim=c(10,20)) +
  blank.plot 
```

## 15 kV Unipolar 20x5

### 2-Parameter Nonlinear Asymptotic Regression

#### Depth

```{r}
drm_depth_asymp_4 <- drm(
  depth ~ application_num,
  data = working_dat,
  subset = (voltage == 15 &
              pulse_seq == "20x5" & 
              waveform == "unipolar"),
  fct = AR.2(names = c("Upper Limit", "Steepness"))
)
```

##### Model Selection 

```{r}
mselect(
  drm_depth_asymp_4,
  list(
    L.3(), #L.3 lower limit at dose 0 is fixed at 0
    LL.3(), #LL.3 lower limit at dose 0 is fixed at 0
    LL2.3(), #LL2.3 lower limit at dose 0 is fixed at 0
    AR.2(), #AR.2 lower limit at dose 0 is fixed at 0
    MM.2(), #MM.2 lower limit at dose = 0 is fixed at 0
  ),
  linreg = TRUE
)
```

```{r}
summary(drm_depth_asymp_4)
```

```{r}
working_dat |>
  filter(voltage == 15, pulse_seq == "20x5", waveform == "unipolar") |>
  ggplot(data = _, aes(x = application_num, y = depth)) +
  geom_smooth(method = drm,
              method.args = list(fct = AR.2()),
              se = F) +
  geom_jitter(width = 0.2, size = 4, na.rm = T) + 
  scale_x_continuous(
    labels = unique(working_dat$application_num),
    breaks = unique(working_dat$application_num)
  ) +
  ggtitle("15kV 20x5 Unipolar") +
  xlab("Applications") +
  ylab("Depth (mm)") +
  # labs(caption = "Caption") + 
  coord_cartesian(ylim=c(5,20))+
  blank.plot +
  theme(axis.text=element_text(size=40, face = "bold"),
        axis.title=element_text(size=24,face="bold")) 

ggsave("15kv20x5uni_depth.png",
       path = "figures",
       scale = 1,
       width = 20,
       height = 28,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

#### Diameter

```{r}
drm_diameter_asymp_4 <- drm(
  diameter ~ application_num,
  data = working_dat,
  subset = (voltage == 15 &
              pulse_seq == "20x5" & 
              waveform == "unipolar"),
  fct = AR.2(names = c("Upper Limit", "Steepness"))
)
```

```{r}
summary(drm_diameter_asymp_4)
```

```{r}
working_dat |>
  filter(voltage == 15, pulse_seq == "20x5", waveform == "unipolar") |>
  ggplot(data = _, aes(x = application_num, y = diameter)) +
  geom_smooth(method = drm,
              method.args = list(fct = AR.2()),
              se = F) +
  geom_jitter(width = 0.2, size = 4, na.rm = T) + 
  scale_x_continuous(
    labels = unique(working_dat$application_num),
    breaks = unique(working_dat$application_num)
  ) +
  ggtitle("15kV 20x5 Unipolar") +
  xlab("Applications") +
  ylab("Diameter (mm)") +
  # labs(caption = "Caption") +
  coord_cartesian(ylim=c(20,60)) +
  blank.plot +
  theme(axis.text=element_text(size=40, face = "bold"),
        axis.title=element_text(size=24,face="bold")) 

ggsave("15kv20x5uni_diameter.png",
       path = "figures",
       scale = 1,
       width = 20,
       height = 28,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

#### Depth/Diameter Ratio

```{r}
rlm_depth_ratio_4 <- rlm(
  depth_per_width ~ application_num,
  data = working_dat,
  subset = (voltage == 15 &
              pulse_seq == "20x5" & 
              waveform == "unipolar"))
```

```{r}
summary(rlm_depth_ratio_4)
f.robftest(rlm_depth_ratio_4,"application_num")
```

```{r}
working_dat |>
  filter(voltage == 15,
              pulse_seq == "20x5",
              waveform == "unipolar") |>
ggplot(data = _, aes(x = application_num, y = depth_per_width)) +
  geom_jitter(width = 0.2, size = 4, na.rm = T) + 
  geom_smooth(method = "rlm", se = T) +
  scale_x_continuous(labels = unique(working_dat$application_num),
                     breaks = unique(working_dat$application_num)) +
  ggtitle("10kV 20x5 Bipolar") +
  xlab("Applications") +
  ylab("Depth per Diameter") +
  # labs(caption = "Caption") +
#  coord_cartesian(ylim=c(10,20)) +
  blank.plot 
```

## Summary Plot

#### Diameter

```{r}
working_dat |>
  ggplot(data = _, aes(x = application_num, y = diameter)) +
  geom_smooth(method = drm,
              method.args = list(fct = AR.2()),
              se = F) +
  geom_jitter(size = 4, aes(color = interaction(voltage,waveform,pulse_seq)), na.rm = T) +
  facet_wrap( ~ voltage + waveform + pulse_seq, nrow = 1) +
  scale_x_continuous(labels = working_dat$application_num,
                     breaks = working_dat$application_num) +
  scale_color_discrete(guide = "none") +
  coord_cartesian(ylim=c(0,60)) +
#  ggtitle("Title") +
  xlab("Applications") +
  ylab("Diameter (mm)") +
#  # labs(caption = "Caption") +
  blank.plot 
#  theme(axis.text=element_text(size=30, face = "bold"),
#        axis.title=element_text(size=24,face="bold")) 

#ggsave("drm_applications_vs_diameter.png",
#       path = "figures",
#       scale = 1,
#       width = 50,
#       height = 25,
#       units = "cm",
#       dpi = "retina",
#       limitsize = F)
```

#### Depth

```{r}
working_dat |>
  ggplot(data = _, aes(x = application_num, y = depth)) +
  geom_smooth(method = drm,
              method.args = list(fct = AR.2()),
              se = F) +
  geom_jitter(size = 4, aes(color = interaction(voltage,waveform,pulse_seq)), na.rm = T) +
  facet_wrap( ~ voltage + waveform + pulse_seq, nrow = 1) +
  scale_x_continuous(labels = working_dat$application_num,
                     breaks = working_dat$application_num) +
  scale_color_discrete(guide = "none") +
  coord_cartesian(ylim=c(0,25)) +
#  ggtitle("Title") +
  xlab("Applications") +
  ylab("Depth (mm)") +
#  # labs(caption = "Caption") +
  blank.plot 
#  theme(axis.text=element_text(size=30, face = "bold"),
#        axis.title=element_text(size=24,face="bold")) 

#ggsave("drm_applications_vs_depth.png",
#       path = "figures",
#       scale = 1,
#       width = 50,
#       height = 25,
#       units = "cm",
#       dpi = "retina",
#       limitsize = F)
```

#### Depth/Diameter Ratio Regression Plot

```{r}
working_dat |>
  ggplot(data = _, aes(
    x = application_num,
    y = depth_per_width
  )) +
  geom_jitter(size = 4, aes(color = interaction(voltage,waveform,pulse_seq)), na.rm = T) +
  geom_smooth(method = "rlm",
              se = F) +
  facet_wrap( ~ voltage + waveform + pulse_seq, nrow = 1) +
  scale_x_continuous(labels = working_dat$application_num,
                     breaks = working_dat$application_num) +
  scale_color_discrete(guide = "none") +
  ggtitle("Title") +
  xlab("Applications") +
  ylab("Depth/Diameter Ratio") +
  # labs(caption = "Caption") +
  blank.plot +
  theme(axis.text=element_text(size=30, face = "bold"),
        axis.title=element_text(size=24,face="bold")) 

ggsave("depth_diameter_ratio_dr.png",
       path = "figures",
       scale = 1,
       width = 50,
       height = 25,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

# Version and Package Details

```{r echo=F}
paste(version$version.string, version$nickname)
paste0(
  "RStudio Version ",
  rstudioapi::versionInfo()$version,
  " ",
  rstudioapi::versionInfo()$release_name
) # RStudio Version
subset(data.frame(sessioninfo::package_info()),
       attached == TRUE,
       c(package, loadedversion)) # Loaded Packages
```

# When were these files last rewritten?

```{r}
date()
```
